+++ R/bayesopt_mpcl.R
@@ -16,7 +16,7 @@
 #'   The [bbotk::OptimInstanceBatchSingleCrit] to be optimized.
 #' @param init_design_size (`NULL` | `integer(1)`)\cr
 #'   Size of the initial design.
-#'   If `NULL` and the [bbotk::Archive] contains no evaluations, \code{4 * d} is used with \code{d} being the
+#'   If `NULL` and the [bbotk::ArchiveBatch] contains no evaluations, \code{4 * d} is used with \code{d} being the
 #'   dimensionality of the search space.
 #'   Points are generated via a Sobol sequence.
 #' @param surrogate ([Surrogate])\cr
@@ -42,7 +42,7 @@
 #' @note
 #' * The `acq_function$surrogate`, even if already populated, will always be overwritten by the `surrogate`.
 #' * The `acq_optimizer$acq_function`, even if already populated, will always be overwritten by `acq_function`.
-#' * The `surrogate$archive`, even if already populated, will always be overwritten by the [bbotk::Archive] of the [bbotk::OptimInstanceBatchSingleCrit].
+#' * The `surrogate$archive`, even if already populated, will always be overwritten by the [bbotk::ArchiveBatch] of the [bbotk::OptimInstanceBatchSingleCrit].
 #' * To make use of parallel evaluations in the case of `q > 1, the objective
 #'   function of the [bbotk::OptimInstanceBatchSingleCrit] must be implemented accordingly.
 #'