+++ R/bayesopt_parego.R
@@ -15,7 +15,7 @@
 #'   The [bbotk::OptimInstanceBatchMultiCrit] to be optimized.
 #' @param init_design_size (`NULL` | `integer(1)`)\cr
 #'   Size of the initial design.
-#'   If `NULL` and the [bbotk::Archive] contains no evaluations, \code{4 * d} is used with \code{d} being the
+#'   If `NULL` and the [bbotk::ArchiveBatch] contains no evaluations, \code{4 * d} is used with \code{d} being the
 #'   dimensionality of the search space.
 #'   Points are generated via a Sobol sequence.
 #' @param surrogate ([SurrogateLearner])\cr
@@ -44,9 +44,9 @@
 #' @note
 #' * The `acq_function$surrogate`, even if already populated, will always be overwritten by the `surrogate`.
 #' * The `acq_optimizer$acq_function`, even if already populated, will always be overwritten by `acq_function`.
-#' * The `surrogate$archive`, even if already populated, will always be overwritten by the [bbotk::Archive] of the [bbotk::OptimInstanceBatchMultiCrit].
+#' * The `surrogate$archive`, even if already populated, will always be overwritten by the [bbotk::ArchiveBatch] of the [bbotk::OptimInstanceBatchMultiCrit].
 #' * The scalarizations of the objective function values are stored as the `y_scal` column in the
-#'   [bbotk::Archive] of the [bbotk::OptimInstanceBatchMultiCrit].
+#'   [bbotk::ArchiveBatch] of the [bbotk::OptimInstanceBatchMultiCrit].
 #' * To make use of parallel evaluations in the case of `q > 1, the objective
 #'   function of the [bbotk::OptimInstanceBatchMultiCrit] must be implemented accordingly.
 #'