+++ tests/testthat/helper.R
@@ -201,6 +201,19 @@ expect_acqfunction = function(acqf) {
   expect_man_exists(acqf$man)
 }
 
+expect_rush_reset = function(rush, type = "kill") {
+  processes = rush$processes
+  rush$reset(type = type)
+  expect_list(rush$connector$command(c("KEYS", "*")), len = 0)
+  walk(processes, function(p) p$kill())
+}
+
+flush_redis = function() {
+  config = redux::redis_config()
+  r = redux::hiredis(config)
+  r$FLUSHDB()
+}
+
 sortnames = function(x) {
   if (!is.null(names(x))) {
     x = x[order(names(x), decreasing = TRUE)]