+++ NEWS.md
@@ -1,5 +1,11 @@
 # mlr3mbo (development version)
 
+* refactor: refactored `SurrogateLearner` and `SurrogateLearnerCollection` to allow updating on an asynchronous `Archive`
+* feat: added experimental `OptimizerAsyncMbo`, `OptimizerADBO`, `TunerAsyncMbo`, and `TunerADBO` that allow for asynchronous optimization
+* feat: added `AcqFunctionStochasticCB` and `AcqFunctionStochasticEI` that are useful for asynchronous optimization
+* doc: minor changes to highlight differences between batch and asynchronous objects related to asynchronous support
+* refactor: `AcqFunction`s and `AcqOptimizer` gained a `reset()` method.
+
 # mlr3mbo 0.2.6
 
 * refactor: Extract internal tuned values in instance.